// laptop_shop_admin/src/app/core/service/model/product.model.ts
export class ProductModel{
  id: number;
  code: string;
  name: string;
  categoryCode: string;
  producerCode: string;
  quantity: number;
  price: number;
  content: string;
  synopsis: string;
  screenSize: string;
  resolution: string;
  os: string;
  cpu: string;
  gpu: string;
  ram: string;
  rom: string;
  batteryCapacity: string;
  weight: string;
  networdConnect: string;
  sale: string;
  imageUrl: string;
  imageUrl2: string;
  imageUrl3: string;
  guarantee: number;
}